new modules; fix localStorage bug
