+++ webpack.config.js
@@ -1,0 +1,23 @@
+// webpack.config.js
+module.exports = {
+  mode: "development",
+  entry: "./src/Map_JS.js",
+  output: {
+    filename: "main.js",
+    publicPath: "dist"
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+          options: {
+            presets: ["@babel/preset-env"]
+          }
+        }
+      }
+    ]
+  }
+};
